Refactor predict and run to share code
