+++ pkg/image/build.go
@@ -49,3 +49,25 @@ func Build(cfg *config.Config, dir, imageName string) error {
 	}
 	return nil
 }
+
+func BuildBase(cfg *config.Config, dir string) (string, error) {
+	// TODO: better image management so we don't eat up disk space
+	// https://github.com/replicate/cog/issues/80
+	imageName := config.BaseDockerImageName(dir)
+
+	console.Info("Building Docker image from environment in cog.yaml...")
+	generator := dockerfile.NewGenerator(cfg, dir)
+	defer func() {
+		if err := generator.Cleanup(); err != nil {
+			console.Warnf("Error cleaning up Dockerfile generator: %s", err)
+		}
+	}()
+	dockerfileContents, err := generator.GenerateBase()
+	if err != nil {
+		return "", fmt.Errorf("Failed to generate Dockerfile: %w", err)
+	}
+	if err := docker.Build(dir, dockerfileContents, imageName); err != nil {
+		return "", fmt.Errorf("Failed to build Docker image: %w", err)
+	}
+	return imageName, nil
+}
